make build directory for bundled files move exports to variables and functions
