+++ .gitignore
@@ -57,3 +57,6 @@ react-native-reanimated-tests.js
 
 # eclipse
 *.settings**
+
+# plugin
+!plugin/build
