+++ plugin/package.json
@@ -10,7 +10,8 @@
   "scripts": {
     "prepare": "yarn generate",
     "generate": "yarn tsc && yarn bundle",
-    "bundle": "yarn esbuild src/plugin.js --bundle --external:@babel --external:convert-source-map --external:fs --log-level=warning --platform=node --sourcemap=linked --outfile=index.js && rm src/*.js src/*.js.map",
+    "clean": "rm src/*.js src/*.js.map",
+    "bundle": "yarn esbuild src/plugin.js --bundle --external:@babel --external:convert-source-map --external:fs --log-level=warning --platform=node --sourcemap=linked --outfile=build/plugin.js && yarn clean",
     "format": "yarn prettier --write --list-different **/*.ts",
     "lint": "yarn eslint --max-warnings=0 --ext .ts . && yarn prettier --check **/*.ts",
     "type:check": "yarn tsc --noEmit"
