+++ plugin/src/makeWorklet.ts
@@ -6,8 +6,36 @@ import { transformSync } from '@babel/core';
 import * as fs from 'fs';
 import * as convertSourceMap from 'convert-source-map';
 import { ReanimatedPluginPass } from './commonInterfaces';
-import { shouldGenerateSourceMap, hash, isRelease } from './commonFunctions';
+import { isRelease } from './commonFunctions';
 import { globals } from './commonObjects';
+
+function hash(str: string): number {
+  let i = str.length;
+  let hash1 = 5381;
+  let hash2 = 52711;
+
+  while (i--) {
+    const char = str.charCodeAt(i);
+    hash1 = (hash1 * 33) ^ char;
+    hash2 = (hash2 * 33) ^ char;
+  }
+
+  return (hash1 >>> 0) * 4096 + (hash2 >>> 0);
+}
+
+function shouldGenerateSourceMap() {
+  if (isRelease()) {
+    return false;
+  }
+
+  if (process.env.REANIMATED_PLUGIN_TESTS === 'jest') {
+    // We want to detect this, so we can disable source maps (because they break
+    // snapshot tests with jest).
+    return false;
+  }
+
+  return true;
+}
 
 function buildWorkletString(
   t: typeof BabelCore.types,
@@ -191,7 +219,7 @@ function makeWorkletName(
   return 'anonymous'; // fallback for ArrowFunctionExpression and unnamed FunctionExpression
 }
 
-function makeWorklet(
+export function makeWorklet(
   t: typeof BabelCore.types,
   fun: BabelCore.NodePath<
     | BabelTypes.FunctionDeclaration
@@ -457,5 +485,3 @@ function makeWorklet(
 
   return newFun;
 }
-
-export { makeWorklet };
